Changed chest and title to svg, reordered Header.tsx imports, renamed schedule-background.svg
