+++ apps/site/src/app/schedule/Schedule.tsx
@@ -3,7 +3,7 @@ import Image from "next/image";
 import { getSchedule } from "./getSchedule";
 import ScheduleTimer from "./ScheduleTimer/ScheduleTimer";
 
-import background from "@/assets/images/resources-background.svg";
+import background from "@/assets/images/schedule-background.svg";
 
 import styles from "./Schedule.module.scss";
